% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap2palette.R
\name{snap2palette}
\alias{snap2palette}
\title{import your own snap and make a palette}
\usage{
snap2palette(path, n, plot_picture = FALSE, plot_palette = TRUE)
}
\arguments{
\item{path}{path to picture in png form}

\item{n}{number of colours in outputted palette}

\item{plot_picture}{plot picture segmented into palette colours, defaults to FALSE (TRUE or FALSE)}

\item{plot_palette}{plot resulting palette, defaults to TRUE}
}
\value{
new palette
}
\description{
import your own snap and make a palette
}
\references{
https://www.r-bloggers.com/color-quantization-in-r/
}
