% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R, R/snapalette.R
\docType{package}
\name{snapalette}
\alias{snapalette}
\alias{snapalette-package}
\title{A palette generator from my holiday snaps
Based on the wesanderson package}
\usage{
snapalette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:}

\item{n}{Number of colors desired.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.
@importFrom graphics rgb rect par image text}
}
\value{
A vector of colours.
}
\description{
list of palettes from my holiday snaps
}
\details{
list of palettes from my holiday snaps
}
\keyword{colors}
