% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{snapalettes}
\alias{snapalettes}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 19.}
\usage{
snapalettes
}
\description{
Use \code{\link{snapalette}} to construct palettes of desired length.
}
\keyword{datasets}
